リファクタリング(役割ごとにファイル分割)
